requirements.txt
box2d-py
dataclasses-json
gym
gym[atari]
matplotlib
numpy
opencv-python
pandas
pygame
sklearn
tb-nightly
torch
torchvision

# Needed for CS224
gensim
nltk
